Files names to snake case
