+++ embedchain/__init__.py
@@ -2,9 +2,8 @@ import importlib.metadata
 
 __version__ = importlib.metadata.version(__package__ or __name__)
 
-from embedchain.apps.App import App  # noqa: F401
-from embedchain.apps.CustomApp import CustomApp  # noqa: F401
-from embedchain.apps.Llama2App import Llama2App  # noqa: F401
-from embedchain.apps.OpenSourceApp import OpenSourceApp  # noqa: F401
-from embedchain.apps.PersonApp import (PersonApp,  # noqa: F401
-                                       PersonOpenSourceApp)
+from embedchain.apps.app import App  # noqa: F401
+from embedchain.apps.custom_app import CustomApp  # noqa: F401
+from embedchain.apps.llama_2_app import Llama2App  # noqa: F401
+from embedchain.apps.open_source_app import OpenSourceApp  # noqa: F401
+from embedchain.apps.person_app import PersonApp, PersonOpenSourceApp  # noqa: F401
