+++ embedchain/apps/person_app.py
@@ -1,11 +1,10 @@
 from string import Template
 
-from embedchain.apps.App import App
-from embedchain.apps.OpenSourceApp import OpenSourceApp
+from embedchain.apps.app import App
+from embedchain.apps.open_source_app import OpenSourceApp
 from embedchain.config import ChatConfig, QueryConfig
-from embedchain.config.apps.BaseAppConfig import BaseAppConfig
-from embedchain.config.QueryConfig import (DEFAULT_PROMPT,
-                                           DEFAULT_PROMPT_WITH_HISTORY)
+from embedchain.config.apps.base_app_config import BaseAppConfig
+from embedchain.config.query_config import DEFAULT_PROMPT, DEFAULT_PROMPT_WITH_HISTORY
 
 
 class EmbedChainPersonApp:
