+++ embedchain/chunkers/youtube_video.py
@@ -3,7 +3,7 @@ from typing import Optional
 from langchain.text_splitter import RecursiveCharacterTextSplitter
 
 from embedchain.chunkers.base_chunker import BaseChunker
-from embedchain.config.AddConfig import ChunkerConfig
+from embedchain.config.add_config import ChunkerConfig
 
 
 class YoutubeVideoChunker(BaseChunker):
