+++ embedchain/config/__init__.py
@@ -1,9 +1,8 @@
-from .AddConfig import AddConfig, ChunkerConfig  # noqa: F401
-from .apps.AppConfig import AppConfig  # noqa: F401
-from .apps.CustomAppConfig import CustomAppConfig  # noqa: F401
-from .apps.OpenSourceAppConfig import OpenSourceAppConfig  # noqa: F401
-from .BaseConfig import BaseConfig  # noqa: F401
-from .ChatConfig import ChatConfig  # noqa: F401
-from .QueryConfig import QueryConfig  # noqa: F401
-from .vectordbs.ElasticsearchDBConfig import \
-    ElasticsearchDBConfig  # noqa: F401
+from .add_config import AddConfig, ChunkerConfig  # noqa: F401
+from .apps.app_config import AppConfig  # noqa: F401
+from .apps.custom_app_config import CustomAppConfig  # noqa: F401
+from .apps.open_source_app_config import OpenSourceAppConfig  # noqa: F401
+from .base_config import BaseConfig  # noqa: F401
+from .chat_config import ChatConfig  # noqa: F401
+from .query_config import QueryConfig  # noqa: F401
+from .vectordbs.elasticsearch_db_config import ElasticsearchDBConfig  # noqa: F401
