--- embedchain/models/VectorDatabases.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class VectorDatabases(Enum):
-    CHROMADB = "CHROMADB"
-    ELASTICSEARCH = "ELASTICSEARCH"
